Refactor agent and incident repositories to remove unused methods and streamline queries
- Removed `GetAll` method from `repo_agents` and `repo_incidents` to simplify data retrieval.
- Updated `Create` method in `repo_agents` to exclude unnecessary fields.
- Added `StatsByServiceID` method in `repo_incidents` for incident statistics retrieval.
- Introduced `GetByServiceID` method in `repo_service_states` for fetching service state by ID.
- Enhanced `Update` method in `repo_service_states` to support field masking for updates.
- Updated service statistics calculation logic to improve accuracy.
- Adjusted data types in DTOs for consistency and accuracy.
- Generated new TypeScript models for frontend integration.

+++ internal/web/dto.go
@@ -18,7 +18,7 @@ type DashboardStats struct {
 	Protocols        map[models.ServiceProtocolType]int `json:"protocols"`
 	ActiveIncidents  int                                `json:"active_incidents" example:"2"`
 	AvgResponseTime  int64                              `json:"avg_response_time" example:"150"`
-	TotalChecks      int                                `json:"total_checks" example:"1000"`
+	TotalChecks      int64                              `json:"total_checks" example:"1000"`
 	UptimePercentage float64                            `json:"uptime_percentage" example:"95.5"`
 	LastCheckTime    *time.Time                         `json:"last_check_time"`
 	ChecksPerMinute  int                                `json:"checks_per_minute" example:"60"`
